+++ src/components/chart/timeline-chart/const.ts
@@ -1,0 +1,18 @@
+export interface TimeLineData {
+  start: Date;
+  end: Date;
+  label: string | null;
+  color?: string;
+}
+
+declare module "chart.js" {
+  interface ChartTypeRegistry {
+    timeline: {
+      chartOptions: BarControllerChartOptions;
+      datasetOptions: BarControllerDatasetOptions;
+      defaultDataPoint: TimeLineData;
+      parsedDataType: any;
+      scales: "timeline";
+    };
+  }
+}
